以目录区分版本

+++ mysql-rsync/mysql-exec.sh
@@ -14,9 +14,16 @@ function_update()
 {
     if [[ -f ${FILE_PATH}/update.sql ]]; then
         command -v mysql >/dev/null 2>&1 ||  (apt-get update && apt-get install -y mysql-client)
-        mysql -h"${CONFIG_MYSQL_HOST}" -u"${CONFIG_MYSQL_USER}" -p"${CONFIG_MYSQL_PASSWORD}" -e "CREATE DATABASE IF NOT EXISTS ${CONFIG_MYSQL_DATABASE} DEFAULT CHARACTER SET utf8 DEFAULT COLLATE utf8_general_ci"
-        mysql -h"${CONFIG_MYSQL_HOST}" -u"${CONFIG_MYSQL_USER}" -p"${CONFIG_MYSQL_PASSWORD}" -D${CONFIG_MYSQL_DATABASE} < ${FILE_PATH}/update.sql
-        rm -f ${FILE_PATH}/update.sql
+
+        SQLHASH=`md5sum ${FILE_PATH}/update.sql|cut -d ' ' -f1`
+        if [ ! -f ${FILE_PATH}/.update.history.log ]; then
+            touch ${FILE_PATH}/.update.history.log
+        fi
+        if [ -z `cat ${FILE_PATH}/.update.history.log|grep $SQLHASH` ]; then
+            mysql -h"${CONFIG_MYSQL_HOST}" -u"${CONFIG_MYSQL_USER}" -p"${CONFIG_MYSQL_PASSWORD}" -e "CREATE DATABASE IF NOT EXISTS ${CONFIG_MYSQL_DATABASE} DEFAULT CHARACTER SET utf8 DEFAULT COLLATE utf8_general_ci"
+            mysql -h"${CONFIG_MYSQL_HOST}" -u"${CONFIG_MYSQL_USER}" -p"${CONFIG_MYSQL_PASSWORD}" -D${CONFIG_MYSQL_DATABASE} < ${FILE_PATH}/update.sql
+            rm -f ${FILE_PATH}/update.sql
+        fi
     else
         echo 'There is no sql file to be executed'
     fi
